pyproject.toml
[project]
name = "synth-data-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diskcache>=5.6.3",
    "faker>=37.11.0",
    "genson>=1.3.0",
    "httpx>=0.28.1",
    "hypothesis>=6.140.3",
    "jinja2>=3.1.6",
    "jsonpath-ng>=1.7.0",
    "jsonschema>=4.25.1",
    "markdown-it-py>=4.0.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "presidio-analyzer>=2.2.360",
    "exrex>=0.11.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.12.0",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "statsmodels>=0.14.5",
    "typer>=0.19.2",
]
